dataset_management/requirements.txt
cycler==0.10.0
Django==2.2.9
django-picklefield==2.0
djangorestframework==3.11.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
XlsxWriter==1.2.7
